requirements.txt
Django
django-solo
numpy
